LayoutTests/http/tests/misc/form-blob-challenge-expected.txt

ALERT: Response: Uploaded blob data: blobdata
